Remove old files and add more docs to the typechecker
