+++ compiler/hash-typecheck/src/infer.rs
@@ -1,15 +1,8 @@
-mod expressions {
+//! Contains functions to traverse the AST and add types to it, while checking it for correctness.
+//!
+//! @@Incomplete(kontheocharis): this is not yet implemented, but basic unification comes first.
 
-}
-
-mod patterns {
-
-}
-
-mod types {
-
-}
-
-mod visitor {
-
-}
+mod expressions {}
+mod patterns {}
+mod types {}
+mod visitor {}
