+++ compiler/hash-typecheck/src/storage/state.rs
@@ -6,17 +6,19 @@ use std::cell::Cell;
 
 /// Keeps track of various information about the current state of the typecheker while traversing
 /// and checking the AST.
+///
+/// @@Volatile: this will probably change a lot when the AST traversing is implemented.
 #[derive(Debug)]
 pub struct TypecheckState {
     /// Whether or not the typecheker is in a `loop` AST block.
     pub in_loop: Cell<bool>,
     /// Whether the current function has returned explicitly at least once.
     pub ret_once: Cell<bool>,
-    /// The return kind of the function being currently checked (if any).
+    /// The return type of the function being currently checked (if any).
     pub func_ret_ty: Cell<Option<TyId>>,
     /// The current source being typechecked.
     pub current_source: Cell<SourceId>,
-    /// A hint about the kind of the pattern in the let statement that is being checked.
+    /// A hint about the type of the pattern in the let statement that is being checked.
     pub let_pattern_hint: Cell<Option<TyId>>,
 }
